Factor Charlatan into interface and ui_kit (something epic)
